Sort of text clipping

+++ src/UI/ECS/Components/RenderingComponents.hpp
@@ -19,7 +19,7 @@ struct Color
 struct QuadRendererComponent
 {
   Color color;
-  Vector4i borderRadius = {0, 0, 0, 0};
+  Vector4f borderRadius = {0, 0, 0, 0};
   Color borderColor = {0.5f, 0.5f, 0.5f, 1.0f};
   Vector4f borderWidths = {0.0f, 0.0f, 0.0f, 0.0f};
   bool clipContents = true;
